+++ nemo/collections/asr/modules/audio_preprocessing.py
@@ -16,13 +16,17 @@ import math
 import random
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
-from typing import Any, Optional
+from typing import Any, Dict, Optional, Tuple
 
 import torch
 from packaging import version
 
 from nemo.collections.asr.parts.numba.spec_augment import SpecAugmentNumba, spec_augment_launch_heuristics
-from nemo.collections.asr.parts.preprocessing.features import FilterbankFeatures, FilterbankFeaturesTA
+from nemo.collections.asr.parts.preprocessing.features import (
+    FilterbankFeatures,
+    FilterbankFeaturesTA,
+    make_seq_mask_like,
+)
 from nemo.collections.asr.parts.submodules.spectr_augment import SpecAugment, SpecCutout
 from nemo.core.classes import Exportable, NeuralModule, typecheck
 from nemo.core.neural_types import (
@@ -51,6 +55,8 @@ except ModuleNotFoundError:
 
 __all__ = [
     'AudioToMelSpectrogramPreprocessor',
+    'AudioToSpectrogram',
+    'SpectrogramToAudio',
     'AudioToMFCCPreprocessor',
     'SpectrogramAugmentation',
     'MaskedPatchAugmentation',
@@ -682,6 +688,200 @@ class CropOrPadSpectrogramAugmentation(NeuralModule):
     @classmethod
     def restore_from(cls, restore_path: str):
         pass
+
+
+class AudioToSpectrogram(NeuralModule):
+    """Transform a batch of input multi-channel signals into a batch of
+    STFT-based spectrograms.
+
+    Args:
+        fft_length: length of FFT
+        hop_length: length of hops/shifts of the sliding window
+        power: exponent for magnitude spectrogram. Default `None` will
+               return a complex-valued spectrogram
+    """
+
+    def __init__(self, fft_length: int, hop_length: int, power: Optional[float] = None):
+        if not HAVE_TORCHAUDIO:
+            logging.error('Could not import torchaudio. Some features might not work.')
+
+            raise ModuleNotFoundError(
+                "torchaudio is not installed but is necessary to instantiate a {self.__class__.__name__}"
+            )
+
+        super().__init__()
+
+        # For now, assume FFT length is divisible by two
+        if fft_length % 2 != 0:
+            raise ValueError(f'fft_length = {fft_length} must be divisible by 2')
+
+        self.stft = torchaudio.transforms.Spectrogram(
+            n_fft=fft_length, hop_length=hop_length, power=power, pad_mode='constant'
+        )
+
+        # number of subbands
+        self.F = fft_length // 2 + 1
+
+    @property
+    def num_subbands(self) -> int:
+        return self.F
+
+    @property
+    def input_types(self) -> Dict[str, NeuralType]:
+        """Returns definitions of module output ports.
+        """
+        return {
+            "input": NeuralType(('B', 'C', 'T'), AudioSignal()),
+            "input_length": NeuralType(('B',), LengthsType()),
+        }
+
+    @property
+    def output_types(self) -> Dict[str, NeuralType]:
+        """Returns definitions of module output ports.
+        """
+        return {
+            "output": NeuralType(('B', 'C', 'D', 'T'), SpectrogramType()),
+            "output_length": NeuralType(('B',), LengthsType()),
+        }
+
+    @typecheck()
+    def forward(self, input: torch.Tensor, input_length: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
+        """Convert a batch of C-channel input signals
+        into a batch of complex-valued spectrograms.
+
+        Args:
+            input: Time-domain input signal with C channels, shape (B, C, T)
+            input_length: Length of valid entries along the time dimension, shape (B,)
+
+        Returns:
+            Output spectrogram with F subbands and N time frames, shape (B, C, F, N)
+            and output length with shape (B,).
+        """
+        output_length = self.get_output_length(input_length=input_length)
+
+        B, T = input.size(0), input.size(-1)
+        input = input.view(B, -1, T)
+
+        # STFT output (B, C, F, N)
+        with torch.cuda.amp.autocast(enabled=False):
+            output = self.stft(input.float())
+
+        # Mask padded frames
+        length_mask: torch.Tensor = make_seq_mask_like(
+            lengths=output_length, like=output, time_dim=-1, valid_ones=False
+        )
+        output = output.masked_fill(length_mask, 0.0)
+
+        return output, output_length
+
+    def get_output_length(self, input_length: torch.Tensor) -> torch.Tensor:
+        """Get length of valid frames for the output.
+
+        Args:
+            input_length: number of valid samples, shape (B,)
+
+        Returns:
+            Number of valid frames, shape (B,)
+        """
+        output_length = input_length.div(self.stft.hop_length, rounding_mode='floor').add(1).long()
+        return output_length
+
+
+class SpectrogramToAudio(NeuralModule):
+    """Transform a batch of input multi-channel spectrograms into a batch of
+    time-domain multi-channel signals.
+
+    Args:
+        fft_length: length of FFT
+        hop_length: length of hops/shifts of the sliding window
+        power: exponent for magnitude spectrogram. Default `None` will
+               return a complex-valued spectrogram
+    """
+
+    def __init__(self, fft_length: int, hop_length: int):
+        if not HAVE_TORCHAUDIO:
+            logging.error('Could not import torchaudio. Some features might not work.')
+
+            raise ModuleNotFoundError(
+                "torchaudio is not installed but is necessary to instantiate a {self.__class__.__name__}"
+            )
+
+        super().__init__()
+
+        # For now, assume FFT length is divisible by two
+        if fft_length % 2 != 0:
+            raise ValueError(f'fft_length = {fft_length} must be divisible by 2')
+
+        self.istft = torchaudio.transforms.InverseSpectrogram(
+            n_fft=fft_length, hop_length=hop_length, pad_mode='constant'
+        )
+
+        self.F = fft_length // 2 + 1
+
+    @property
+    def num_subbands(self) -> int:
+        return self.F
+
+    @property
+    def input_types(self) -> Dict[str, NeuralType]:
+        """Returns definitions of module output ports.
+        """
+        return {
+            "input": NeuralType(('B', 'C', 'D', 'T'), SpectrogramType()),
+            "input_length": NeuralType(('B',), LengthsType()),
+        }
+
+    @property
+    def output_types(self) -> Dict[str, NeuralType]:
+        """Returns definitions of module output ports.
+        """
+        return {
+            "output": NeuralType(('B', 'C', 'T'), AudioSignal()),
+            "output_length": NeuralType(('B',), LengthsType()),
+        }
+
+    @typecheck()
+    def forward(self, input: torch.Tensor, input_length: torch.Tensor) -> torch.Tensor:
+        """Convert input complex-valued spectrogram to a time-domain
+        signal. Multi-channel IO is supported.
+
+        Args:
+            input: Input spectrogram for C channels, shape (B, C, F, N)
+            input_length: Length of valid entries along the time dimension, shape (B,)
+
+        Returns:
+            Time-domain signal with T time-domain samples and C channels, (B, C, T)
+            and output length with shape (B,).
+        """
+        output_length = self.get_output_length(input_length=input_length)
+
+        B, F, N = input.size(0), input.size(-2), input.size(-1)
+        assert F == self.F, f'Number of subbands F={F} not matching self.F={self.F}'
+        input = input.view(B, -1, F, N)
+
+        # iSTFT output (B, C, T)
+        with torch.cuda.amp.autocast(enabled=False):
+            output = self.istft(input.cfloat())
+
+        # Mask padded samples
+        length_mask: torch.Tensor = make_seq_mask_like(
+            lengths=output_length, like=output, time_dim=-1, valid_ones=False
+        )
+        output = output.masked_fill(length_mask, 0.0)
+
+        return output, output_length
+
+    def get_output_length(self, input_length: torch.Tensor) -> torch.Tensor:
+        """Get length of valid samples for the output.
+
+        Args:
+            input_length: number of valid frames, shape (B,)
+
+        Returns:
+            Number of valid samples, shape (B,)
+        """
+        output_length = input_length.sub(1).mul(self.istft.hop_length).long()
+        return output_length
 
 
 @dataclass
